pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.2"
dependencies = [
    "fastapi>=0.120.2",
    "ipython>=9.6.0",
    "langchain-community>=0.3.31",
    "langchain-milvus>=0.2.1",
    "langchain-openai>=0.3.35",
    "langchain-pymupdf4llm>=0.4.1",
    "oci-cli>=3.69.0",
    "pillow>=12.0.0",
    "pymilvus[milvus-lite]>=2.6.3",
    "python-dotenv>=1.2.1",
    "rapidocr-onnxruntime>=1.4.4",
    "schema>=0.7.8",
    "utils>=1.0.2",
]
